Move formatting classes to serialisation package
