+++ src/main/java/uk/gov/register/serialisation/SerialisationFormatter.java
@@ -1,4 +1,4 @@
-package uk.gov.register.util;
+package uk.gov.register.serialisation;
 
 import uk.gov.register.core.Entry;
 import uk.gov.register.core.Item;
