+++ src/main/java/uk/gov/register/util/ArchiveCreator.java
@@ -7,6 +7,7 @@ import io.dropwizard.jackson.Jackson;
 import uk.gov.register.core.Entry;
 import uk.gov.register.core.Item;
 import uk.gov.register.core.RegisterDetail;
+import uk.gov.register.serialisation.SerialisationFormatter;
 
 import javax.ws.rs.core.StreamingOutput;
 import java.io.Closeable;
