+++ src/test/java/uk/gov/register/util/ArchiveCreatorTest.java
@@ -10,6 +10,7 @@ import uk.gov.register.core.Entry;
 import uk.gov.register.core.Item;
 import uk.gov.register.core.RegisterData;
 import uk.gov.register.core.RegisterDetail;
+import uk.gov.register.serialisation.SerialisationFormatter;
 
 import javax.ws.rs.core.StreamingOutput;
 import java.io.ByteArrayInputStream;
